+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@adobe/helix-project-boilerplate",
+  "name": "hsf-commonmoves",
   "private": true,
   "version": "1.0.0",
   "description": "Starter project for Adobe Helix",
@@ -10,14 +10,14 @@
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/adobe/helix-project-boilerplate.git"
+    "url": "git+https://github.com/hlxsites/hsf-commonmoves.git"
   },
   "author": "Adobe",
   "license": "Apache License 2.0",
   "bugs": {
-    "url": "https://github.com/adobe/helix-project-boilerplate/issues"
+    "url": "https://github.com/hlxsites/hsf-commonmoves/issues"
   },
-  "homepage": "https://github.com/adobe/helix-project-boilerplate#readme",
+  "homepage": "https://github.com/hlxsites/hsf-commonmoves#readme",
   "devDependencies": {
     "@babel/core": "7.21.0",
     "@babel/eslint-parser": "7.19.1",
