+++ scripts/scripts.js
@@ -12,7 +12,7 @@ import {
   loadBlocks,
   loadCSS,
   getMetadata,
-} from './lib-franklin.js';
+} from './aem.js';
 
 export const LIVEBY_API = 'https://api.liveby.com/v1/';
 
@@ -23,7 +23,6 @@ export const BREAKPOINTS = {
 };
 
 const LCP_BLOCKS = ['hero']; // add your LCP blocks to the list
-window.hlx.RUM_GENERATION = 'bhhs-commonmoves'; // add your RUM generation information here
 
 export function preloadHeroImage(picture) {
   const src = [...picture.querySelectorAll('source')]
@@ -41,7 +40,17 @@ export function preloadHeroImage(picture) {
   link.setAttribute('type', src.getAttribute('type'));
   document.head.append(link);
 }
-
+/**
+ * load fonts.css and set a session storage flag
+ */
+async function loadFonts() {
+  await loadCSS(`${window.hlx.codeBasePath}/styles/fonts.css`);
+  try {
+    if (!window.location.hostname.includes('localhost')) sessionStorage.setItem('fonts-loaded', 'true');
+  } catch (e) {
+    // do nothing
+  }
+}
 /**
  * Builds hero block and prepends to main in a new section.
  * @param {Element} main The container element
@@ -217,7 +226,16 @@ async function loadEager(doc) {
   const main = doc.querySelector('main');
   if (main) {
     decorateMain(main);
+    document.body.classList.add('appear');
     await waitForLCP(LCP_BLOCKS);
+  }
+  try {
+    /* if desktop (proxy for fast connection) or fonts already loaded, load fonts.css */
+    if (window.innerWidth >= 900 || sessionStorage.getItem('fonts-loaded')) {
+      loadFonts();
+    }
+  } catch (e) {
+    // do nothing
   }
 }
 
@@ -260,6 +278,7 @@ async function loadLazy(doc) {
   loadFooter(doc.querySelector('footer'));
 
   loadCSS(`${window.hlx.codeBasePath}/styles/lazy-styles.css`);
+  loadFonts();
   addFavIcon(`${window.hlx.codeBasePath}/styles/favicon.svg`);
   sampleRUM('lazy');
   sampleRUM.observe(main.querySelectorAll('div[data-block-name]'));
